End to end evaluation parse / compile / eval tests.
Note: in order to best surface issues within individual components,
the examples directory should be incorporated into the parser,
compiler, and runtime tests.

Note: here's a slight usability issue with a missing 'type'.
The absent type value is a legitimate use case for Open API Schema,
but does make it significantly more painful to identify when a
type should have been specified when it was not.