Move custom type name specification into schema metadata.
The environment and the template schema both specify custom type definitions
which may be backed by protobufs under the covers. The custom type definitions
should be specifiable in the same way to permit custom functions which accept
policy instance types as input.

Additionally, it is important to be able to support K8s CRD schemas as input
types directly into the environment, so the added alignment with Open API Schema
provides ease of use improvements with K8s as well.

There are also some refactors to share more code between OpenAPISchema and
DeclType compilation